+++ src/main/java/ru/practicum/shareit/user/repository/UserRepository.java
@@ -8,13 +8,13 @@ public interface UserRepository {
 
     User create(User user);
 
-    User get(Integer id);
+    User get(Long id);
 
     List<User> getAll();
 
-    User update(Integer id,User user);
+    User update(Long id,User user);
 
-    void delete(Integer id);
+    void delete(Long id);
 
 
 }
